Move multiplication tables exercise
