+++ README.md
@@ -4,7 +4,6 @@ Ce dépôt contient un programme `exercices` permettant d'exécuter différents 
 
 Le menu propose plusieurs exercices :
 - `anglais_hello_world` affiche `hello world!`
-- `math_addition` affiche `1+1=2`
 - `hist_test` affiche `La préhistoire c'est ce qu'il y a avant l'histoire.`
 - `physique_changements_etats` propose une leçon et un quiz sur les changements d'état de la matière.
 - `math_tables_multiplication` propose une leçon et un quiz interactif pour pratiquer une table de multiplication.
